+++ changelog.md
@@ -1,2 +1,2 @@
 ### Entangled 1.3.20
-- Initial release of Entangled for Minecraft 1.21.6 & 1.21.7
+- Initial release of Entangled for Minecraft 1.21.9 & 1.21.10